Move iptables steps to dedicated jobs so they run when the deploy fails
by having a different job for iptables, we should work around the issue
where the concourse deployment seems to fail because we lose our worker
so the iptables job never runs and dns breaks until we re-run the whole
deployment job. By getting the production deployment we can be somewhat
sure that we'll notice the deployment finish asynchronously, then we'll
run these tasks before too long. There is a better solution - we should
really recover from the worker disappearing and resume tailing the task
logs - but that's a fix for another day